+++ src/main/java/online/kingdomkeys/kingdomkeys/entity/mob/BaseKHEntity.java
@@ -6,6 +6,7 @@ import net.minecraft.network.syncher.EntityDataAccessor;
 import net.minecraft.network.syncher.EntityDataSerializers;
 import net.minecraft.network.syncher.SynchedEntityData;
 import net.minecraft.server.level.ServerLevel;
+import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.DifficultyInstance;
 import net.minecraft.world.damagesource.DamageSource;
@@ -46,6 +47,11 @@ public class BaseKHEntity extends Monster implements IKHMob {
 
 	public boolean isBoss() {
 		return getKHMobType() == MobType.BOSS;
+	}
+
+	@Nullable
+	public SoundEvent arrivalSound() {
+		return null;
 	}
 
 	public int getDeathSequence() {