+++ src/main/java/online/kingdomkeys/kingdomkeys/entity/mob/LightTrainingOrbEntity.java
@@ -2,12 +2,14 @@ package online.kingdomkeys.kingdomkeys.entity.mob;
 
 import net.minecraft.core.particles.ParticleOptions;
 import net.minecraft.core.particles.ParticleTypes;
+import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.monster.Monster;
 import net.minecraft.world.level.Level;
+import online.kingdomkeys.kingdomkeys.client.sound.ModSounds;
 
 public class LightTrainingOrbEntity extends TrainingOrbEntity {
 	public LightTrainingOrbEntity(EntityType<? extends Monster> type, Level level) {
@@ -23,6 +25,11 @@ public class LightTrainingOrbEntity extends TrainingOrbEntity {
 				.add(Attributes.ATTACK_DAMAGE, 1.0D)
 				.add(Attributes.ATTACK_KNOCKBACK, 0.2D)
 				.add(Attributes.KNOCKBACK_RESISTANCE, 0.0D);
+	}
+
+	@Override
+	public SoundEvent arrivalSound() {
+		return ModSounds.lightOrbSpawn.get();
 	}
 
 	@Override